view: first pass at the viewer
we can figure out pass pipelines; i.e.:
```
Pipeline:
  ┠─ IREEImportPublic
  ┠─ ImportMLProgram
  ┠─ SanitizeModuleNames
  ┠─ mlir::iree_compiler::IREE::ABI::WrapEntryPointsPass
  ┠─ Inliner
  ┃  ┠─ Canonicalizer
  ┃  ┠─ Canonicalizer
  ┃  ┠─ Canonicalizer
  ┃  ┖─ Canonicalizer
  ┠─ Canonicalizer
  ┠─ CSE
  ┠─ SymbolDCE
  ┠─ DemoteF64ToF32
  ┠─ ConvertConv2D1x1ConvToMatmul
  ┠─ DetachElementwiseFromNamedOps
  ┠─ VerifyInputLegality
  ┠─ LinalgNamedOpConversion
  ┠─ ExpandTensorShapes
  ┠─ FixedPointIterator
  ┃  ┠─ SimplifyGlobalAccesses
  ┃  ┠─ ApplyPatterns
  ┃  ┠─ FoldGlobals
  ┃  ┠─ Canonicalizer
  ┃  ┖─ CSE
  ┠─ PadTensorToTensorInsertSlice
  ┠─ ConvertElementwiseToLinalg
  ┠─ LinalgFoldUnitExtentDims
  ┠─ InterchangeGenericOps
  ┠─ ResolveShapedTypeResultDims
  ┠─ Canonicalizer
<snipped>
```

Next up is wiring up delta, etc.

I want to be able to use `delta` for the diffs and since it's a binary
only crate this means actually invoking the binary.

Unfortunately I don't think there's an easy way to, like, invoke a
binary and have it live in a tui widget.

So, rather than reinvent `tmux` and all the VT-100 parsing and
forwarding and TTY shenanigans that that entails... I'm just going to
have the viewer set up a `tmux` session. Two panes, one with a `tui` app
the other with a thing that just runs commands (delta to start but this
can be customizable). A little bit of IPC and a parent process to shut
down tmux and we should be good, I think.